Ensure compatibility with async sock.bind() in node 0.10+
Per <http://nodejs.org/api/dgram.html#dgram_event_listening>: "the behavior of
dgram.Socket#bind() has changed in v0.10 and is always asynchronous now."
This means that setting the broadcast flag has to be deferred to a callback to
ensure it does not execute before the socket is bound. Setting the flag in a
'listening' event handler is compatible with both old and new versions of node.

Change-Id: I6c1928fda13110f326fb025048788033338a9167