Allow user to specify custom metric filters / preprocessors
Not all metrics that are appropriate for Graphite are also appropriate for
Ganglia. This patch adds an optional configuration option, "gangliaFilters",
that the user may set to an array of module paths. Each referenced module is
expected to export a "filter" function which takes a metric and either modifies
it in some way, returns it unmodified, or returns a false value to exclude it
from Ganglia reporting.

Change-Id: Ieca6e923b14278b243bd1ca80611f2cea7f751b3