Ensure metadata is emitted every send_metadata_interval
Metrics may be sent infrequently, but gmetad expects metric metadata to be sent
at regular intervals. This patch ensures that metadata is emitted every sixty
seconds or fewer. This can be tweaked by setting 'sendMetadataInterval' in the
configuration.

Also included: fix to default tmax. If StatsD's flushInterval exceeds the
default tmax, set the default tmax to the flushInterval.

Change-Id: I6639bdd37ce4106333b7e9defb41ce2b3e50f075